Написан механизм отображения данных

+++ back/sanitation_web_api/domain/SanitationDto.cs
@@ -3,7 +3,7 @@
     public class SanitationDto
     {
         public int Id { get; set; }
-        public required decimal[] Values { get; set; }
+        public required float[] Values { get; set; }
         public int Year { get; set; }
         public int SourceId { get; set; }
         public int DistrictId { get; set; }
